Only create db if it doesn't exist
As a best practice, the database should be created by an admin account the nut influxdb user account should only get write access to the database. To handle this scenario, first check to see if the database exists, if it does, continue. If the database doesn't exist, try to create it (this will succeed if authentication is disabled and/or if the user has the necessary permissions). It will raise an exception if the nut influxdb user doesn't have the necessary permissions to create the database.

Also group the functions at the top to make reading the code a bit easier.